feat 优化Thread Dump，新增线程统计信息接口
(cherry picked from commit b1a3ef70ea4c8c3018c50864e35f62c1761618a0)

+++ core/src/main/java/org/beifengtz/jvmm/core/Unsafe.java
@@ -56,7 +56,7 @@ public final class Unsafe {
      * @return Thread实例
      * @throws Exception 调用异常
      */
-    public static Thread getThread(int threadId) throws Exception {
+    public static Thread getThread(long threadId) throws Exception {
         for (Thread thread : getThreads()) {
             if (thread.getId() == threadId) {
                 return thread;
